refactor: split `useNotes.js` into smaller hooks.
fix(ui): fix paddings for textareas of note header and note content for better visual harmony.
