+++ styles/MainContent.css
@@ -420,7 +420,7 @@
   display: flex;
   align-items: center;
   gap: 16px;
-  padding: 32px 16px 0px 32px;
+  padding: 32px 16px 16px 32px;
   flex-shrink: 0;
 }
 
@@ -485,7 +485,7 @@
   line-height: 1.8;
   font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto",
     sans-serif;
-  padding: 16px 32px;
+  padding: 0 32px 160px 32px;
   transition: all 0.3s ease;
   overflow-y: auto;
   overflow-x: hidden;
